Revert "openfile started"
This reverts commit 3013535d37b602842d7cfd231c51ecd92ec4a4c6.

+++ GUI/cgp_functions.c
@@ -696,7 +696,7 @@ int get_num_generations(void)
     return NUM_GENERATIONS;
 }
 
-int get_penalty(void)
+double get_penalty(void)
 {
     return PENALTY;
 }
@@ -755,7 +755,7 @@ void set_num_generations(int a)
     NUM_GENERATIONS = a;
 }
 
-void set_penalty(int a)
+void set_penalty(double a)
 {
     PENALTY = a;
 }
